Remove dependency of RC namespace in rc-mail-messages
