+++ packages/rocketchat-mail-messages/client/startup.js
@@ -1,10 +1,11 @@
-import { RocketChat } from 'meteor/rocketchat:lib';
+import { AdminBox } from 'meteor/rocketchat:ui-utils';
+import { hasAllPermission } from 'meteor/rocketchat:authorization';
 
-RocketChat.AdminBox.addOption({
+AdminBox.addOption({
 	href: 'mailer',
 	i18nLabel: 'Mailer',
 	icon: 'mail',
 	permissionGranted() {
-		return RocketChat.authz.hasAllPermission('access-mailer');
+		return hasAllPermission('access-mailer');
 	},
 });
