+++ packages/rocketchat-mail-messages/client/views/mailer.js
@@ -1,12 +1,13 @@
 import { Meteor } from 'meteor/meteor';
 import { Template } from 'meteor/templating';
 import { TAPi18n } from 'meteor/tap:i18n';
-import { RocketChat, handleError } from 'meteor/rocketchat:lib';
+import { settings } from 'meteor/rocketchat:settings';
+import { handleError } from 'meteor/rocketchat:utils';
 import toastr from 'toastr';
 
 Template.mailer.helpers({
 	fromEmail() {
-		return RocketChat.settings.get('From_Email');
+		return settings.get('From_Email');
 	},
 });
