+++ packages/rocketchat-mail-messages/server/functions/unsubscribe.js
@@ -1,8 +1,8 @@
-import { RocketChat } from 'meteor/rocketchat:lib';
+import { Users } from 'meteor/rocketchat:models';
 
 export const unsubscribe = function(_id, createdAt) {
 	if (_id && createdAt) {
-		return RocketChat.models.Users.rocketMailUnsubscribe(_id, createdAt) === 1;
+		return Users.rocketMailUnsubscribe(_id, createdAt) === 1;
 	}
 	return false;
 };
