+++ packages/rocketchat-mail-messages/server/index.js
@@ -1,5 +1,4 @@
 import './startup';
-import './models/Users';
 import './methods/sendMail';
 import './methods/unsubscribe';
 import { Mailer } from './lib/Mailer';
