+++ packages/rocketchat-mail-messages/server/methods/sendMail.js
@@ -1,6 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 import { Mailer } from '../lib/Mailer';
-import { RocketChat } from 'meteor/rocketchat:lib';
+import { hasRole } from 'meteor/rocketchat:authorization';
 
 Meteor.methods({
 	'Mailer.sendMail'(from, subject, body, dryrun, query) {
@@ -10,7 +10,7 @@ Meteor.methods({
 				method: 'Mailer.sendMail',
 			});
 		}
-		if (RocketChat.authz.hasRole(userId, 'admin') !== true) {
+		if (hasRole(userId, 'admin') !== true) {
 			throw new Meteor.Error('error-not-allowed', 'Not allowed', {
 				method: 'Mailer.sendMail',
 			});
