+++ packages/rocketchat-mail-messages/server/startup.js
@@ -1,8 +1,8 @@
 import { Meteor } from 'meteor/meteor';
-import { RocketChat } from 'meteor/rocketchat:lib';
+import { Permissions } from 'meteor/rocketchat:models';
 
 Meteor.startup(function() {
-	return RocketChat.models.Permissions.upsert('access-mailer', {
+	return Permissions.upsert('access-mailer', {
 		$setOnInsert: {
 			_id: 'access-mailer',
 			roles: ['admin'],
